fix(ratelimiting): dont spam the slack api with requests
previously the @slack_update_delay was not being resolved correctly,
leading to a problem with the check that verifies the age of the tmp
file where the previous request is stored. this check would always fail,
and as a result of the ! would always succeed causing a request to
always be sent.